f2f-main/f2f/Backend/bsas_supplychain-main/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
psycopg2-binary>=2.9.0
qrcode[pil]>=7.0
